LayoutTests/imported/w3c/web-platform-tests/mediacapture-streams/enumerateDevices-with-navigation.https-expected.txt


FAIL enumerateDevices with navigation assert_equals: Promise state while inactive expected "pending" but got "resolved"
